requirements_for_tests.txt
-r requirements.txt
nose
PyHamcrest
Mock